GRHB-120: * refactors
