--- mobile/src/common/enums/group/group.ts
@@ -1,4 +1,0 @@
-export {
-  GroupValidationMessage,
-  GroupValidationRule,
-} from 'guruhub-shared/common/enums/group/group';
