+++ mobile/src/common/enums/navigation/app-screen-name.enum.ts
@@ -6,7 +6,7 @@ enum AppScreenName {
   BILLING = 'Billing',
   SETTINGS = 'Settings',
   UAM = 'User Access Managment',
-  UAM_GROUP_CREATION = 'Create group',
+  UAM_GROUPS_CREATE = 'Create group',
 }
 
 export { AppScreenName };
