--- mobile/src/common/types/user/user-group-creation.dto.type.ts
@@ -1,10 +1,0 @@
-import { ReactElement } from 'react';
-
-type UsersGroupCreationDto = {
-  checkbox: ReactElement;
-  fullName: string;
-  email: string;
-  id: number;
-};
-
-export { type UsersGroupCreationDto };
