+++ mobile/src/common/types/user/user.ts
@@ -1,4 +1,3 @@
-export { type UsersGroupCreationDto } from './user-group-creation.dto.type';
 export {
   type UsersByIdResponseDto,
   type UsersDeleteRequestParamsDto,
