+++ mobile/src/components/uam-groups-create/common/enums/permission-table-accessor.enum.ts
