+++ mobile/src/components/uam-groups-create/common/types/permissions-table-row.type.ts
@@ -1,4 +1,7 @@
+import { ReactElement } from 'react';
+
 type PermissionsTableRow = {
+  checkbox: ReactElement;
   name: string;
   id: number;
 };
