+++ mobile/src/components/uam-groups-create/components/conponents.ts
@@ -1,0 +1,2 @@
+export { GroupsTable } from './permissions-table/permissions-table';
+export { UsersTable } from './users-table/users-table';
