+++ mobile/src/components/uam-groups-create/helpers/group-creation-get-user-columns/group-creation-get-user-columns.helper.ts
@@ -1,6 +1,6 @@
 import { TableColumn } from '~/common/types/ui/table-column.type';
-import { UserTableAccessor } from '~/components/uam/components/uam-group-creation/common/enums/enums';
-import { UsersTableRow } from '~/components/uam/components/uam-group-creation/common/types/types';
+import { UserTableAccessor } from '~/components/uam-groups-create/common/enums/enums';
+import { UsersTableRow } from '~/components/uam-groups-create/common/types/types';
 
 const getUserColumns = (): TableColumn<UsersTableRow>[] => {
   return [
