+++ mobile/src/components/uam-groups-create/helpers/helpers.ts
