--- mobile/src/components/uam/components/uam-group-creation/components/conponents.ts
@@ -1,2 +1,0 @@
-export { GroupCreationPermissionsTable } from './permissions-table/permissions-table';
-export { GroupCreationUsersTable } from './users-table/users-table';
