+++ mobile/src/navigation/app/app.navigation.tsx
@@ -6,7 +6,7 @@ import {
   AppNavigationParamList,
   DrawerNavigationItem,
 } from '~/common/types/types';
-import { UamGroupCreation } from '~/components/uam/components/components';
+import { UAMGroupsCreate } from '~/components/uam/components/components';
 import { useAppSelector, useMemo } from '~/hooks/hooks';
 import { getAllowedScreens } from '~/navigation/app/helpers/helpers';
 
@@ -45,8 +45,8 @@ const App: FC = () => {
       })}
 
       <Drawer.Screen
-        name={AppScreenName.UAM_GROUP_CREATION}
-        component={UamGroupCreation}
+        name={AppScreenName.UAM_GROUPS_CREATE}
+        component={UAMGroupsCreate}
       />
     </Drawer.Navigator>
   );
